updated_requirements.txt
apache-airflow==2.6.0
cx_Oracle==8.3.0
apache-airflow-providers-oracle==3.3.0
openmetadata-ingestion==1.6.12.0
sqlalchemy==1.4.49
alembic==1.12.0
pydantic==1.10.8
python-dateutil==2.8.2
